feat: Portal tunneling to remote endpoints (#44)
Add an option to the portal server to define a remote host to tunnel for
a list of remote endpoint types. The remote endpoint types are set to
the APIC and User service endpoints for now. The portal client queries
the list of remote endpoint types and their associated hosts and then
uses that list to set the tunnel string remote host. The forward port
grpc method is updated to take a remote endpoint type. The callers to
the forward port grpc method in the Kurtosis repo will be updated in a
subsequent PR.

Tested with a bastion + remote backend setup with this version of the
portal running on the bastion and on the user machine.

Design doc:
https://www.notion.so/kurtosistech/Portal-with-tunneling-to-a-remote-endpoint-df7a7b450ced48bd8f05ffaf9e6bcb1b