fix: Forward port wait until ready option (#46)
CLI commands like `context switch` send requests to the forwarded engine
port right after forwarding the port. Those commands sometimes fail
because the port is not forwarded yet. The forward port operation is
asynchronous. The portal client relies on the Chisel client async Start
method. AFAIK there is no option today in the Chisel client to be told
when the connection is ready. This change adds a function to wait for
the TCP/UDP connection to be ready. This is driven by a new boolean flag
`WaitUntilReady` (defaulting to false) in the ForwardPort RPC call
arguments.

Release 0.0.6 will contain this change.